requirements.txt
# pytorch imports
pytorch
torchvision

# other imports
notebook
yapf